Move LearningSessionImpl to background task runner.
In order to reduce the load on the main thread, this CL wraps
LearningSessionImpl in a SequenceBound on a background thread.

This is likely not the correct threading model long-term, but
LearningHelper shouldn't own a LearningSessionImpl anyway.  The
goal is to see if some metrics regression in page load time
go away if we move everything off the main thread.

Change-Id: Ide6b1e266f95306b680d3d2aebfdc41eadeb7ba6
Reviewed-on: https://chromium-review.googlesource.com/c/1441192
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#627327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5f4d6e60cc9c653e9f6a08a354f42d30157cbd8