Delete *MediaPlayerManager and RemoteMediaPlayerBridge
This CL removes RendererMediaPlayerManager and its browser counterpart,
the RemoteMediaPlayerManager. It also removes the
RemoteMediaPlayerBridge, which is tigthly coupled ot the RMPM.

The BrowserMediaPlayerManager cannot be removed yet, as it is coupled
with MediaPlayerRenderer. Getting rid of BMPM and the MediaPlayerAndroid
interface will be taken care of in a follow up CL.

Bug: 575276, 517102, 570711, 790766, 926595
Change-Id: I5296e7611c104cae35384c64f6d73ea1cc3d5a29
Reviewed-on: https://chromium-review.googlesource.com/c/1444553
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#627591}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b2d059ed8ce988fb2fb10d77ac800e302020caf3