media: Add MediaLog on selected DecryptingDemuxerStream
Currently when DecryptingDemuxerStream is selected, we only have
"video_dds = true" in MediaLog, which isn't obvious for readers. This CL
adds an explicit log event to make it clear.

Unit test is not updated because it's not trivial to set when we expect this to
be logged. For example, even with encrypted stream and a valid Decryptor, we
don't expect this log when decoder initialization failed.

Test: Manually tested. See note above.
Change-Id: If2149cd73c51916f1aa55e6f073ed2f73d65f28a
Reviewed-on: https://chromium-review.googlesource.com/c/1457432
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#630033}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 842e677ad1e08f4364b8138424ab23812d4e96c1