media/gpu/v4l2vda: revamp EGL fences handling
When being returned a CAPTURE buffer, the VDA must make sure that all
potential running operations on it are completed before reusing it. To
that end an EGL fence is attached to buffers when they are returned so
the VDA refrains from using them until it is signaled.

The way we were doing this so far was to return buffers in wait of being
signaled to the free buffers list and make the decoder thread wait on the
fence prior to queueing a buffer. Since the buffer queue is FIFO, we are
always waiting on the very next buffer that will be available.

This works, but has the effect of making the decoder thread wait for
potentially long periods of time, during which nothing else can be done.
During that time, other events may also happen on the client side: for
instance, request for destructing the VDA. This leads to a threads dance
that is error-prone and needs to be explicitly managed.

This CL revamps the way we wait on EGL fences by integrating them into
the decoding pipeline: buffers waiting to be signaled are stored in a
set, and are not returned to the free buffers list until their fence is
signaled. That way, the decoder thread never blocks on a fence and can do
other tasks while waiting for a buffer.

It also simplifies the code by not having to care whether a destroy event
has happened while we were waiting, making things overall safer. Cherry
on the cake, we also remove one member of OutputRecord, bringing it one
step closer to its demotion.

It appears that with the previous code, the VDA could starve on buffers
in some situations. With this CL, performance of the Thumbnail unittest
on Hana has increased from ~170 fps to ~260 fps.

Bug: 792790
Test: Run VDA unittest on Hana, Nyan_big and Peach_pit. Also checked
crosvideo.appspot.com on these 3 boards.

Change-Id: Ib56a8d5d5ddc6530f9863ee32579f3412e9ccf5f
Reviewed-on: https://chromium-review.googlesource.com/c/1460210
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#631523}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0f3b891971f64c6c3a191afb1309fddadaaeb1e9