media/gpu/v4l2vda: keep reference to IP VideoFrame sent to client
Since we want the IP's buffers to be managed independently of the VDA,
the IP needs to know when it can reuse one of its buffers that have been
passed to the client. A VideoFrame destruction observer is the mechanism
we are going to use for this, so make the VDA keep a reference of the IP
VideoFrames that it has passed to the client, and erase that reference
once the client returns it.

Bug: 792790
Test: Run VDA unittest on Hana and Peach_pit.

Change-Id: I5d78cc6706eb4a8b9dc6dd037000bcce732d16a5
Reviewed-on: https://chromium-review.googlesource.com/c/1460592
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#632076}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a67c7c34a15a832633e25e3d8422a059a9c1636