Add MediaDrmOriginIdManager to pre-provision origin IDs on Android
To support encrypted media playback on Android when the Internet is not
available (e.g. on a plane), we need to pre-provision some origin IDs
so that provisioning is completed while connected. This adds a new
MediaDrmOriginIdManager class that keeps track of the pre-provisioned
origin IDs in a new preference "media.media_drm_origins".

MediaDrmStorageFactory now uses the new class to get a pre-provisioned
origin ID rather than generating a new origin ID that needs to be
provisioned before it can be used.

BUG=917527
TEST=new unit_tests pass

Change-Id: If4a5af4ba50b0586d5ca6f6f7cb8ea9686f98fa9
Reviewed-on: https://chromium-review.googlesource.com/c/1416398
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#632468}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4703af69f5fac9a4f622730c6fb377e432f1f4fb