media/gpu/test: Remove FlushBeforeResetDone test.
The FlushBeforeResetDone test checks whether the decoder behaves correctly when
triggering a flush while resetting. But this behavior is considered invalid in
the media::VideoDecoder interface:
https://cs.chromium.org/chromium/src/media/base/video_decoder.h?l=114&rcl=5348a2ac7bde8050c8677df8d4b02c9f9df75cdf

We are moving our decoders to the media::VideoDecoder interface. Chrome already
accesses the decoders through this interface by making use of a wrapper, so it's
safe to deprecate this testcase.

TEST=ran new VDA tests on nocturne

BUG=938710

Change-Id: I85da6c860f77a70454dc9421e1bc5e8a83a3e98e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1504593
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#638828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f71f82722c68ed7d5201acbf6658bdd5af349dbc