Single-feature importance measurement.
In order to determine which features are most predictive for a
learning task, this CL adds the beginnings of a random feature
subset selection algorithm.  The idea is that clients will select a
random subset of features, and the learning results will reflect a
model trained on just those features.

Currently, since we're reporting results via UMA, we allow only a
subset of size one for importance measuring.  This makes it easier
to report results on a single histogram.

Change-Id: I33aa9659c13e07376a9ff801dda030f76567fa75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531107
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#642329}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0daff27b6c68cff5b3be706181c0c4727a897277