Set is_encrypted_ in common SetCdm() path
SetCdm() is called from two places
        WMPI constructor, when provided an intial CDM
        SetContentDecryptionModule, when keys are set after load()

Only the second case was triggering the is_encrypted_ and relevant
state to be set.

For WatchTimeReporter this should have only a minor impact. Before this
change, we initially created the reporter with incorrect state whenever
the CDM was provided via the constructor. But this was quickly
remmedied by re-creating in OnEncryptedMediaInitData().

This CL has a bigger impact on a future change to create the
VideoDecodeStatsReporter for recording encrypted stats. Unlike
WatchTimeReporter, this object will only be created in SetCDM and not
OnEncryptedMediaInitData() because it depends on properties of the CDM
for construction.

Bug: 907909
Change-Id: I2e8a41fae89cddee98e7d97b98809eaceae9174e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1522206
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#643076}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d4a0019cd03e786589984fe07d0a5b24cab61929