media: Add UMA to monitor MjpegDecodeAccelerator::Error for VAAPI.
This CL creates the Media.VAJDA.ResponseToClient UMA to monitor the
distribution of responses emitted by the VaapiJpegDecodeAccelerator to
the client for decode requests. This is useful to make sure that changes
to the VAAPI JPEG decoder code don't cause unexpected regressions in the
wild.

Additionally, the description of the existing Media.VAJDA.DecoderFailure
UMA is updated to distinguish it from the new UMA, and ownership is
transferred to myself because the current owner moved to a different
team.

the resulting histogram.

Test: Ran JDA/SimpleDecode and FailureJpeg tests with --v=1 and observed
Bug: 868400
Change-Id: I535e6cb5c283054ed02933f05b70256fc9536e26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1533288
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#643713}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2beb30953bd1cc796e69c4ca88906266713fd705