[Video Capture, ChromeOS] Fix wrong rotation when starting capture with non-zero rotation
When starting video capture with a device that is rotated, e.g. a
convertible in portrait orientation, the rotation of the captured video
intermittently (most of the time) comes out wrong.

The root cause is that rotation events are ignored before capturing is
started but may end up arriving before.

This CL fixes this by storing the rotation in case rotation events are
received while capturing is not started.

Bug: 944215
Change-Id: I1346772341044c92be754d11d17a519397598df9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540459
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Sheng-hao Tsao <shenghao@chromium.org>
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#644783}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74a6f7fbbbd950c50ff300b05193a90ea8e41a1c