blink/cc: Detect images that are completely received initially.
This CL adds support to detect when the encoded data for an image has
been received in its entirety prior to creating the first
DecodingImageGenerator for it. This is to support the use case of
sending the encoded data to a hardware decoder. The motivation is as
follows: hardware decoding requires that all the data has been received.
If the image is received incrementally and we software-decode the
partial data as it comes, there is no point in sending the complete
image to the hardware decoder once it's received because the software
decoder has already done a lot of the work, so it's better to finish the
decoding in software.

A DeferredImageDecoder is shared across the process of incrementally
receiving the image. However, at each step, a new DecodingImageGenerator
is created (which is what cc uses ultimately to decode the image).
Therefore, we keep enough state in DeferredImageDecoder to detect when
we create the first DecodingImageGenerator. If at that point, all the
encoded data has been received, then it's okay to send the image to
hardware decoder. We store this decision in the DecodingImageGenerator.
That way a PaintImage can have an IsEligibleForAcceleratedDecoding()
method which asks the underlying PaintImageGenerator if it's okay to
hardware-decode.

Additionally, a small clean-up is done: the
DecodingImageGenerator::SetCanYUVDecode() method is removed in favor of
passing the YUV decoding decision at construction.

Test: unit tests are added for the new behavior.
Bug: 868400
Change-Id: I396cf08ab007fd98b6fd65d60704ae8d26673aa4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1546327
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Madeleine Barowsky <mbarowsky@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#646411}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ae8e59d489f1fbcf2f018973a0c9769068a07868