media/VJD: Determine supported internal -> image format conversions.
This CL adds support for requesting the result of a JPEG decode in an
arbitrary FOURCC. This will allow us to eventually get the result in
RGBX format. It also allows us to refactor how we handle the case in
which regardless of whether the input JPEG is YUV420 or YUV422, we want
I420 output: this is okay with Intel but not with AMD Stoney Ridge
(i.e., grunt). There was already code to handle this Intel/AMD
discrepancy. However, this relied on the grunt driver not reporting I420
as a supported image format which seems incorrect because it can convert
to IYUV (synonym of I420) in vaGetImage() (see
https://gitlab.freedesktop.org/mesa/mesa/commit/b0a44f10).

This CL centralizes the determination of a suitable output FOURCC in the
new VaapiWrapper::GetJpegDecodeSuitableImageFourCC() static method. The
caller can express a preferred FOURCC. If the conversion is not
supported, this method returns a fallback FOURCC that is supported.

Unit tests are added to make sure that all the combinations of supported
subsamplings and supported image formats are successful: for now, we can
check the decode result against software decoding as long as the output
is either I420 or YUY2/YUYV.

Test: JDA tests pass on grunt, nocturne, and samus.
Bug: 868400
Change-Id: I1a8dee1599469544bbaf6fd40bcd5662f7637fdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1494803
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#646562}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 20615d43fbaef7b0d6921bdcfb5940f3b70e9ff7