VideoSurfaceLayer: add flag to fully disable feature.
Because of the 3-state nature of the feature, having WebView use a specific
flag to fully disable it will be safer for the moment. When the feature will
be on/off, we will be able to use only one flag and avoid the double negation.

Bug: 946822
Change-Id: Ib8cc4e9cd020ee0ce5b81a4506bcb841952d7457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1546075
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#646875}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 28e08a0a224a167105a6119996500de5067a07ae