media/gpu/test: Add parameter to overwrite default video test metadata file.
When running the video decoder tests on a video file, it's assumed there is an
accompanying metadata file in the same folder with the ".json" suffix. This CL
allows overriding the default metadata location so it doesn't need to be
present in the same folder.

This CL also makes some of the loading failure error messages a bit more
explicit.

TEST=ran new VDA (perf) tests on nocturne

BUG=879065

Change-Id: I6f7cc5db0b61d28ff8356d7eda6a473fc9950fab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552392
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#649445}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cd76481fcd85a38cb235fc4451b4c08ff791dd43