Report confusion matrices separated by training set weight.
When analyzing the learning results, it's helpful to see the effect
of the amount of training data on the quality of the predictions.
However, DistributionReporter only reports the results as one
confusion matrix per LearningTask.  This merges the predictions from
models with potentially wildly different amounts of training data,
making it hard to see how the model's ramp up time differs from its
steady state performance.

This CL lets DistributionReporter report confusion matrices that are
split up by the total weight of the training data used to train the
model that made the prediction.  The UMA x-axis is divided into
buckets, numbered 0 to 15.  The bucket number is provided in the
label name for the ByTrainingWeight histogram.

Which bucket is used is based on the total weight of the training
set used to train the model.  By default, bucket 0 will report the
confusion matrix for models trained on [0, 9] examples.  Bucket 1
will report [10, 19] examples, etc.  Bucket 9 will report [90,
99].  The overflow bucket, bucket 10, will report 100 and up.  This
was chosen to match the default maximum training set size of 100,
if each example in the set has the default weight of 1.

This CL also modifies LearningHelper to report the weighted task
results in buckets of 50000 frames each, such as [0-49999] for all
tasks that weight examples by the number of frames.

Last, this CL renames the histograms to:

Media.Learning.BinaryThreshold.Aggregate.$LEARNING_TASK_NAME for the
total, aggregate confusion matrix from all predictions (i.e. this is
the same as it reported before).

Media.Learning.BinaryThreshold.ByTrainingWeight.$LEARNING_TASK_NAME
for the bucketed-by-weight matrices described above.

Change-Id: I75953747b62ebefd9d7fa4f427489f683e608117
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1557077
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#649796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7cfca514d6d1a254d3f12e4e88858c18ec29757a