VideoFrame: add a DECODE_COMPLETE_TIMESTAMP to metadata and add it to trace
This CL adds a Metadata DECODE_COMPLETE_TIMESTAMP to the VideoFrame, and
fills it for non-rtc Video deco (at decoder stream traits) and for
remote webrtc video (so e.g. video capture is excluded).

This timestamp is then compared to base::TimeTicks::Now() to add a
TRACE in VideoFrameSubmitter indicating the delay between decode
and CompositorFrameSink entry point (and there there's another
indication of the time to the OS compositor, so those two can be
added to give an idea of the delay between decode and raster). The
media pipeline is then added as a step previous to VFSubmitter's
graphics delay measurements.

Examples:
1) crosvideo vp9 [1]
2) appr.tc session [2] (local capture has no bufferin information).

[1] https://i.imgur.com/28oA1Ft.png (https://imgur.com/a/jtxkLWm)
[2] https://i.imgur.com/6XP24ny.png (https://imgur.com/a/TAjqI4b)

Bug: 950183
Change-Id: I421840d32c815e1d7eed8e99f392a3c36ab2f06e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1560384
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#650233}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 927e4efd1bb30903db23cdba688733ea70f9b52d