media: Add a fuzzertest for the the Jpeg Parser.
A fuzzertest to uncover vulnerabilities in media/filters/jpeg_parser, specifically
for ParseJpegPicture. It uses media/test/data as a seed corpus and
media/test/jpeg.dict as a dictionary.

jpeg.dict is a copy of
https://cs.chromium.org/chromium/src/pdf/pdfium/fuzzers/dicts/pdf_codec_jpeg.dict
which is a part of our repo. It was copied instead of referenced to avoid a media->pdf
dependency.

Bug: 951060
Test: temporarily introduced a CHECK(false) for marker 0xAA in the parser and verified locally that the fuzzer broke the target.
Change-Id: Id6771c49f84b1ed4a058c8e61edbad817b5c383c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1566222
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Madeleine Barowsky <mbarowsky@chromium.org>
Commit-Queue: Gil Dekel <gildekel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#651499}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4924a4161750766180bbfdf5e880986721e24166